Move kSymResolver to ksym module.
